+++ IRR Model/Domain/DTOs/CapitalSchedule.cs
@@ -1,15 +1,15 @@
 ﻿namespace IRR.Domain.DTOs
 {
-    public class CapitalSchedule(int Id, decimal IncrementalCapitalAdded, DateTime date)
+    public class CapitalSchedule(int Id, double IncrementalCapitalAdded, DateTime date)
     {
 
         private readonly int _id = Id;
-        private readonly decimal _incrementalCapitalAdded = IncrementalCapitalAdded;
+        private readonly double _incrementalCapitalAdded = IncrementalCapitalAdded;
         private readonly DateTime _date = date;
 
 
         public int Id => _id;
-        public decimal IncrementalCapitalAdded => _incrementalCapitalAdded;
+        public double IncrementalCapitalAdded => _incrementalCapitalAdded;
 
         public DateTime Date => _date;
