+++ IRR Model/Domain/DTOs/PremiumSchedule.cs
@@ -1,16 +1,18 @@
 ﻿
+using System.Runtime.Serialization;
+
 namespace IRR.Domain.DTOs
 {
     [Serializable]
     public class PremiumSchedule(int RetroProfileId, int RetroProgramId,
         int SPInvestorId,
-        DateTime EarnedDay, decimal UnadjustedPremium,
-        decimal EarnedPremium, int Year)
+        DateTime EarnedDay, double UnadjustedPremium,
+        double EarnedPremium, int Year)
     {
         private readonly int _Year = Year;
         private readonly DateTime _EarnedDay = EarnedDay;
-        private readonly decimal? _UnadjustedPremium = UnadjustedPremium;
-        private readonly decimal? _EarnedPremium = EarnedPremium;
+        private readonly double? _UnadjustedPremium = UnadjustedPremium;
+        private readonly double? _EarnedPremium = EarnedPremium;
         private readonly int _RetroProfileId = RetroProfileId;
         private readonly int _RetroProgramId = RetroProgramId;
         private readonly int _SPInvestorId = SPInvestorId;
@@ -26,8 +28,8 @@ namespace IRR.Domain.DTOs
         public int Year => _Year;
         public DateTime EarnedDay => _EarnedDay;
 
-        public decimal UnadjustedPremium => _UnadjustedPremium ?? 0;
+        public double UnadjustedPremium => _UnadjustedPremium ?? 0;
 
-        public decimal EarnedPremium => _EarnedPremium ?? 0; 
+        public double EarnedPremium => _EarnedPremium ?? 0; 
     }
 }
